% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daum_keyword_R1.R
\name{daum_keyword_R1}
\alias{daum_keyword_R1}
\title{Daum Suggest Search Terms - Search Level 1}
\usage{
daum_keyword_R1(keyword)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function aims to crawl the first level related search terms from the DAUM search engine.
The only argument for this function is "Search Term".
}
\details{
DAUM url: https://www.daum.net/
}
\examples{
daum_keyword_R1("data+scientist")
}
\keyword{daum,}
\keyword{search}
\keyword{terms}
