% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_keyword.R
\name{google_keyword}
\alias{google_keyword}
\title{Google Suggest Search Terms - Search Level 2}
\usage{
google_keyword(keyword)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function aims to crawl the second level related search terms from the Google search engine.
The only argument for this function is "Search Terms".
This function retrieves the google_keyword_R1 function in order to get the related search terms from the intial search term.
It iterates crawling the individual search terms until it consumes all the level 1 search terms.
It refer to the Google Suggest API.
}
\details{
# GOOGLE url: https://www.google.com/
}
\examples{
google_keyword("druking")
}
\keyword{google,}
\keyword{search}
\keyword{terms}
