% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_keyword_R1.R
\name{google_keyword_R1}
\alias{google_keyword_R1}
\title{Google Suggest Search Terms - Search Level 1}
\usage{
google_keyword_R1(keyword)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function aims to crawl the first level related search terms from the Google search engine.
The only argument for this function is "Search Term".
It refer to the Google Suggest API.
}
\details{
GOOGLE url: https://www.google.com/
}
\examples{
google_keyword_R1("korea")
}
\keyword{google,}
\keyword{search}
\keyword{terms}
