% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_list2df.R
\name{google_list2df}
\alias{google_list2df}
\title{Google List to Dataframe}
\usage{
google_list2df(input_list, keyword)
}
\arguments{
\item{input_list}{Google search terms result which must be list datatype.}

\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function convert list to dataframe.
It receives the input from the Google related search terms
in the form of list data structure.
It convert the list datatype into the dataframe.
}
\details{
The return ouput will be used for the subsequent data analysis and visualization
}
\examples{
google_list2df(druking_list, "druking")
}
\keyword{conversion}
\keyword{dataframe,}
\keyword{google,}
\keyword{list}
\keyword{search}
\keyword{terms,}
\keyword{to}
