% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naver_keyword.R
\name{naver_keyword}
\alias{naver_keyword}
\title{Naver Suggest Search Terms - Search Level 2}
\usage{
naver_keyword(keyword)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function aims to crawl the second level related search terms from the NAVER search engine.
The only argument for this function is "Search Terms".
This function retrieves the naver_keyword_R1 function in order to get the related search terms from the intial search term.
It iterates crawling the individual search terms until it consumes all the level 1 search terms.
}
\examples{
naver_keyword("korea")
}
\keyword{naver,}
\keyword{search}
\keyword{terms}
