% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naver_keyword_R1.R
\name{naver_keyword_R1}
\alias{naver_keyword_R1}
\title{Naver Suggest Search Terms - Search Level 1}
\usage{
naver_keyword_R1(keyword)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function aims to crawl the first level related search terms from the NAVER search engine.
The only argument for this function is "Search Term".
}
\details{
NAVER url: https://www.naver.com/
}
\examples{
naver_keyword_R1("data+scientist")
}
\keyword{naver,}
\keyword{search}
\keyword{terms}
