% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_list2df.R
\name{suggest_list2df}
\alias{suggest_list2df}
\title{List to Dataframe from NAVER and DAUM search terms}
\usage{
suggest_list2df(input_list, keyword)
}
\arguments{
\item{input_list}{Naver and Daum search terms result which must be list datatype.}

\item{keyword}{search terms, keywords which supports Korean as well.}
}
\description{
The function convert list to dataframe.
It receives the input from the Daum and Naver related search terms
in the form of list data structure.
It convert the list datatype into the dataframe.
}
\details{
The return ouput will be used for the subsequent data analysis and visualization
}
\examples{
korea_list <- daum_keyword("korea")
suggest_list2df(korea_list, "korea")
}
\keyword{conversion}
\keyword{dataframe,}
\keyword{daum,}
\keyword{list}
\keyword{naver,}
\keyword{search}
\keyword{terms,}
\keyword{to}
