% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_suggest.R
\name{visualize_suggest}
\alias{visualize_suggest}
\title{Visualize Keyword, 1st Search Terms and 2nd Search Terms}
\usage{
visualize_suggest(keyword, search_engine)
}
\arguments{
\item{keyword}{search terms, keywords which supports Korean as well.}

\item{search_engine}{search engine; "daum", "naver", "google"}
}
\description{
The function visualizes the relationships among keyword,
1st search terms and 2nd search terms
The only search term is needed to get visualized artefacts.
}
\examples{
visualize_suggest("korea", "google")
}
\keyword{network,}
\keyword{visualize}
